+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # uwsgiconf changelog
 
-### Unreleased
+### v2.2.0 [2025-09-20]
 * ++ Django Admin. Add 'active' flag for Task (see #15).
 * ++ Django Admin. Add execution duration to task list (see #15).
 * ++ Django Admin. Forced task run for task not registered with uWSGI now issues an error hint (see #15).